+++ base/src/main/java/proguard/analysis/cpa/util/TriFunction.java
@@ -1,0 +1,34 @@
+/*
+ * ProGuardCORE -- library to process Java bytecode.
+ *
+ * Copyright (c) 2002-2022 Guardsquare NV
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package proguard.analysis.cpa.util;
+
+/**
+ * A function taking three parameters.
+ *
+ * @author Dmitry Ivanov
+ */
+@FunctionalInterface
+public interface TriFunction<T, U, V, R>
+{
+
+    /**
+     * Returns the result of the function on given arguments
+     */
+    R apply(T t, U u, V v);
+}