+++ base/src/test/kotlin/proguard/analysis/cpa/HeapOperatorsTest.kt
@@ -31,6 +31,7 @@ import proguard.analysis.cpa.state.MapAbstractStateFactory
 import proguard.classfile.MethodSignature
 import proguard.testutils.ClassPoolBuilder
 import proguard.testutils.JavaSource
+import java.util.Optional
 import kotlin.reflect.full.memberProperties
 import kotlin.reflect.jvm.isAccessible
 
@@ -100,17 +101,17 @@ class HeapOperatorsTest : FreeSpec({
         listOf(
             HashMapAbstractStateFactory.getInstance(),
             DifferentialMapAbstractStateFactory<String, TaintAbstractState> { false },
-            LimitedHashMapAbstractStateFactory { _, _, _ -> false }
+            LimitedHashMapAbstractStateFactory { _, _, _ -> Optional.empty() }
         ).forEach { staticFieldMapAbstractStateFactory ->
             listOf<Pair<MapAbstractStateFactory<Reference, HeapNode<SetAbstractState<Reference>>>, MapAbstractStateFactory<Reference, HeapNode<TaintAbstractState>>>>(
                 Pair(HashMapAbstractStateFactory.getInstance(), HashMapAbstractStateFactory.getInstance()),
                 Pair(DifferentialMapAbstractStateFactory { false }, DifferentialMapAbstractStateFactory { false }),
-                Pair(LimitedHashMapAbstractStateFactory { _, _, _ -> false }, LimitedHashMapAbstractStateFactory { _, _, _ -> false })
+                Pair(LimitedHashMapAbstractStateFactory { _, _, _ -> Optional.empty() }, LimitedHashMapAbstractStateFactory { _, _, _ -> Optional.empty() })
             ).forEach { (principalHeapMapAbstractStateFactory, followerHeapMapAbstractStateFactory) ->
                 listOf<Pair<MapAbstractStateFactory<String, SetAbstractState<Reference>>, MapAbstractStateFactory<String, TaintAbstractState>>>(
                     Pair(HashMapAbstractStateFactory.getInstance(), HashMapAbstractStateFactory.getInstance()),
                     Pair(DifferentialMapAbstractStateFactory { false }, DifferentialMapAbstractStateFactory { false }),
-                    Pair(LimitedHashMapAbstractStateFactory { _, _, _ -> false }, LimitedHashMapAbstractStateFactory { _, _, _ -> false })
+                    Pair(LimitedHashMapAbstractStateFactory { _, _, _ -> Optional.empty() }, LimitedHashMapAbstractStateFactory { _, _, _ -> Optional.empty() })
                 ).forEach { (principalHeapNodeMapAbstractStateFactory, followerHeapNodeMapAbstractStateFactory) ->
                     val testNameSuffix =
                         """