+++ docs/md/releasenotes.md
@@ -4,6 +4,7 @@
 
 - Don't report warnings for missing Kotlin default implementation classes when initializing with `ClassReferenceInitializer`.
 - Only link matching methods in Kotlin file facades with `MethodLinker`.
+- Extend the `LimitedHashMap` parameterization with an element exclusion selector.
 
 ### Bug fixes
 - Fix the reduce operator producing a wrong `JvmAbstractState` for the composite taint analysis.